[Refactor] Split JSON flattener into focused utility modules
