+++ be/src/storage/tablet_reader.cpp
@@ -49,7 +49,8 @@
 #include "storage/tablet_updates.h"
 #include "storage/types.h"
 #include "types/logical_type.h"
-#include "util/json_flattener.h"
+#include "util/json_flat_path.h"
+#include "util/json_path_deriver.h"
 
 namespace starrocks {
