+++ be/test/exprs/flat_json_functions_test.cpp
@@ -39,6 +39,7 @@
 #include "types/json_value.h"
 #include "types/logical_type.h"
 #include "util/json_flattener.h"
+#include "util/json_path_deriver.h"
 
 namespace starrocks {
