+++ be/test/util/json_flattener_test.cpp
@@ -35,6 +35,7 @@
 #include "column/json_column.h"
 #include "column/nullable_column.h"
 #include "column/vectorized_fwd.h"
+#include "common/bloom_filter.h"
 #include "common/config_exec_fwd.h"
 #include "common/config_json_flat_fwd.h"
 #include "common/object_pool.h"
@@ -47,7 +48,7 @@
 #include "gutil/strings/strip.h"
 #include "types/json_value.h"
 #include "types/logical_type.h"
-#include "util/json_flattener.h"
+#include "util/json_path_deriver.h"
 
 namespace starrocks {
